Update gift section wording
I don't really want to suggest people give gifts to charity, since that
sort of contradicts the statement that we don't want gifts. It also
could imply that we are expecting gifts, which can be considered rude.
If people ignore the instructions and give gifts, I would guess many of
them would be money anyway.

I'd also prefer to avoid saying we don't like gifts, since this could
hurt someone's feelings who has given one of us a gift in the past.